refactor: accept unknown chunks in persisted delete predicates
Due to the timing of the "persist" lifecycle action and that delete
predicates might arrive at any time + the fact that we don't wanna hold
transaction locks for too long, we should accept delete predicates for
chunks that are currently "persisting" even though that lifecycle action
might fail.

+++ server/src/db.rs
@@ -1219,8 +1219,14 @@ pub(crate) fn checkpoint_data_from_catalog(catalog: &Catalog) -> CheckpointData 
             };
 
             files.insert(path, m);
+        }
 
-            // capture delete predicates
+        // capture delete predicates
+        // We should only report persisted chunks or chunks that are currently being persisted, because the
+        // preserved catalog does not care about purely in-mem chunks.
+        if matches!(guard.stage(), ChunkStage::Persisted { .. })
+            || guard.is_in_lifecycle(ChunkLifecycleAction::Persisting)
+        {
             for predicate in guard.delete_predicates() {
                 let predicate_ref: &Predicate = predicate.as_ref();
                 let addr = (predicate_ref as *const Predicate) as usize;
